requirements.txt
# Backend
fastapi==0.115.0
uvicorn==0.30.0
python-dotenv==1.0.1
pydantic==2.9.0

# Google APIs
google-api-python-client==2.149.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1

# Telegram Bot
python-telegram-bot==21.6

# AI (respuestas inteligentes)
anthropic==0.39.0

# Utils
httpx==0.27.0
jinja2==3.1.4
aiosqlite==0.20.0
